pyproject.toml
[project]
name = "tron-address-service"
version = "0.1.0"
description = "Microservice for Tron address information"
authors = [
    {name = "Chernevich", email = "kostya_chernevich@mail.ru"},
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "pydantic>=1.8.0",
    "tronpy>=0.2.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "httpx>=0.0.0",
    "mypy>=0.0.0",
    "types-requests>=2.0.0",
    "types-python-dateutil>=0.0.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true